+++ Headers/VideoModel/VideoModel.h
@@ -3,8 +3,8 @@
 #include "Core/Core.h"
 
 struct VideoModel {
-    size_t SegmentDurationInMs;
-    std::vector<size_t> BitRatesInKbps;
+    double SegmentDurationInMs;
+    std::vector<double> BitRatesInKbps;
     std::vector<std::vector<size_t>> SegmentByteCounts;
 
     static VideoModel Import(std::string_view fileName);
@@ -13,7 +13,7 @@ struct VideoModel {
         return SegmentByteCounts.size();
     }
 
-    [[nodiscard]] size_t LengthInMs() const {
-        return SegmentCount() * SegmentDurationInMs;
+    [[nodiscard]] double LengthInMs() const {
+        return static_cast<double>(SegmentCount()) * SegmentDurationInMs;
     }
 };
